# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TripFit API",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "https://api.tripfit.online",
      "description": "Production API Server"
    }
  ],
  "security": [
    {
      "bearer-jwt": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users/google-calendar": {
      "post": {
        "tags": [
          "Google Calendar"
        ],
        "summary": "Google Calendar 연동",
        "description": "목적: Google OAuth authorization code로 Calendar 읽기 권한을 연동한다.\n\n호출 시점: 온보딩·설정 화면에서 Google Calendar 연동 버튼 완료 직후.\n\n전제: TripFit 로그인(JWT) 상태. 앱·웹이 Google OAuth 동의 후 authorization code를 전달한다.\n\n결과: isGoogleCalendarConnected=true로 갱신된 UserSummary. 연동 직후 freeBusy 1회 sync.\n\n주요 에러: GOOGLE_CALENDAR_CONNECT_FAILED — code 교환·Google API 실패\n",
        "operationId": "connect",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectGoogleCalendarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "연동 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseUserSummaryResponse"
                },
                "example": {
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "email": "user@example.com",
                    "firstName": "길동",
                    "lastName": "홍",
                    "nickname": "홍길동",
                    "profileImageUrl": "https://lh3.googleusercontent.com/a/example",
                    "provider": "GOOGLE",
                    "isGoogleCalendarConnected": true,
                    "hasPreSchedule": false,
                    "isAllFree": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "요청 값 검증 실패 (INVALID_INPUT)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "INVALID_INPUT",
                  "message": "입력값이 올바르지 않습니다.",
                  "errors": [
                    {
                      "field": "authorizationCode",
                      "message": "필수 값입니다."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "502": {
            "description": "GOOGLE_CALENDAR_CONNECT_FAILED — authorization code 교환·Google API 호출 실패",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "GOOGLE_CALENDAR_CONNECT_FAILED",
                  "message": "Google Calendar 연동에 실패했습니다."
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Google Calendar"
        ],
        "summary": "Google Calendar 연동 해제",
        "description": "목적: Google Calendar 연동을 의도적으로 해제한다.\n\n호출 시점: 설정·온보딩에서 연동 해제 선택 시.\n\n전제: isGoogleCalendarConnected=true.\n\n결과: revoke(best-effort) 후 credential·busy_day 삭제, flag=false UserSummary. 정기·개별 일정은 유지.\n\n주요 에러: GOOGLE_CALENDAR_NOT_CONNECTED — 미연동 상태에서 해제 요청\n",
        "operationId": "disconnect",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "연동 해제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseUserSummaryResponse"
                },
                "example": {
                  "data": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "email": "user@example.com",
                    "firstName": "길동",
                    "lastName": "홍",
                    "nickname": "홍길동",
                    "profileImageUrl": "https://lh3.googleusercontent.com/a/example",
                    "provider": "GOOGLE",
                    "isGoogleCalendarConnected": false,
                    "hasPreSchedule": false,
                    "isAllFree": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "액세스 토큰 없음·무효(AUTH_INVALID_TOKEN)·만료(AUTH_EXPIRED)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "AUTH_EXPIRED",
                  "message": "액세스 토큰이 만료되었습니다."
                }
              }
            }
          },
          "409": {
            "description": "GOOGLE_CALENDAR_NOT_CONNECTED — 미연동 상태에서 해제 요청",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "code": "GOOGLE_CALENDAR_NOT_CONNECTED",
                  "message": "Google Calendar가 연동되어 있지 않습니다."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "API 에러 응답 envelope",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드",
            "example": "AUTH_INVALID_TOKEN"
          },
          "message": {
            "type": "string",
            "description": "사용자용 에러 메시지",
            "example": "유효하지 않은 토큰입니다."
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "description": "필드 검증 오류 목록 (validation 실패 시)",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "description": "요청 필드 검증 오류",
        "properties": {
          "field": {
            "type": "string",
            "description": "검증 실패 필드명",
            "example": "token"
          },
          "message": {
            "type": "string",
            "description": "검증 실패 메시지",
            "example": "must not be blank"
          }
        }
      },
...
      "ConnectGoogleCalendarRequest": {
        "type": "object",
        "description": "Google Calendar OAuth 연동 요청",
        "properties": {
          "authorizationCode": {
            "type": "string",
            "description": "Google OAuth authorization code (앱·웹이 Google 동의 후 수신)",
            "example": "4/0AeanS...",
            "minLength": 1
          }
        },
        "required": [
          "authorizationCode"
        ]
      },
      "SuccessResponseUserSummaryResponse": {
        "type": "object",
        "description": "API 성공 응답 envelope",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserSummaryResponse",
            "description": "응답 본문"
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 메시지 (성공 시 null)"
          },
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "에러 코드 (성공 시 null)"
          }
        }
      },
      "UserSummaryResponse": {
        "type": "object",
        "description": "사용자 요약. POST /auth/login · GET /auth/me · PATCH /users/onboarding/name · PATCH /users/profile 응답 공통",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "TripFit 사용자 ID (UUID v4)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "description": "소셜 계정 이메일. provider가 제공하지 않으면 null",
            "example": "user@example.com"
          },
          "firstName": {
            "type": [
              "string",
              "null"
            ],
            "description": "사용자가 입력한 이름. 미입력 시 null",
            "example": "길동"
          },
          "lastName": {
            "type": [
              "string",
              "null"
            ],
            "description": "사용자가 입력한 성. 미입력 시 null",
            "example": "홍"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ],
            "description": "소셜 provider 표시명. 프로필 입력 prefill용. 미제공 시 null",
            "example": "홍길동"
          },
          "profileImageUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "프로필 이미지 URL. 현재는 소셜 provider CDN URL. 미제공 시 null",
            "example": "https://lh3.googleusercontent.com/a/example"
          },
          "provider": {
            "type": "string",
            "description": "로그인에 사용한 소셜 제공자",
            "enum": [
              "KAKAO",
              "GOOGLE",
              "APPLE"
            ]
          },
          "isGoogleCalendarConnected": {
            "type": "boolean",
            "description": "Google Calendar OAuth 연동 여부",
            "example": false
          },
          "hasPreSchedule": {
            "type": "boolean",
            "description": "정기 또는 개별 일정이 1건 이상 있는지 (DB 컬럼 없음, 조회 시 계산).\ntrue: 정기 일정 첫 생성 또는 개별 일정 첫 저장.\nfalse: 두 종류 일정 row가 모두 0건.\n일정 CRUD 응답에는 미포함 — GET /auth/me 등 재호출.\n",
            "example": false
          },
          "isAllFree": {
            "type": "boolean",
            "description": "전부 free 선언 여부 (user.is_all_free 저장, 기본 false).\n방 입장 조건: hasPreSchedule OR isAllFree.\ntrue: 일정 없이 Skip+0행으로 create/join·confirm 시 설정.\nfalse: 사용자가 명시적으로 해제하거나, 일정 row가 생기면 hasPreSchedule으로 대체 가능.\n일정 CRUD 응답에는 미포함 — GET /auth/me 등 재호출.\n",
            "example": false
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "type": "http",
        "description": "Authorization: Bearer {accessToken}",
        "name": "bearer-jwt",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}